+++ entrypoint.sh
@@ -3,6 +3,12 @@ set -e
 
 echo -e "\e[34m[Info] Sourcebot version: $SOURCEBOT_VERSION\e[0m"
 
+# If we don't have a PostHog key, then we need to disable telemetry.
+if [ -z "$POSTHOG_PAPIK" ]; then
+    echo -e "\e[33m[Warning] POSTHOG_PAPIK was not set. Setting SOURCEBOT_TELEMETRY_DISABLED.\e[0m"
+    export SOURCEBOT_TELEMETRY_DISABLED=1
+fi
+
 # Issue a info message about telemetry
 if [ ! -z "$SOURCEBOT_TELEMETRY_DISABLED" ]; then
     echo -e "\e[34m[Info] Disabling telemetry since SOURCEBOT_TELEMETRY_DISABLED was set.\e[0m"
@@ -25,7 +31,7 @@ if [ ! -f "$FIRST_RUN_FILE" ]; then
     # (if telemetry is enabled)
     if [ -z "$SOURCEBOT_TELEMETRY_DISABLED" ]; then
         curl -L -s --header "Content-Type: application/json" -d '{
-            "api_key": "'"$POSTHOG_KEY"'",
+            "api_key": "'"$POSTHOG_PAPIK"'",
             "event": "install",
             "distinct_id": "'"$SOURCEBOT_INSTALL_ID"'",
             "properties": {
@@ -43,7 +49,7 @@ else
 
         if [ -z "$SOURCEBOT_TELEMETRY_DISABLED" ]; then
             curl -L -s --header "Content-Type: application/json" -d '{
-                "api_key": "'"$POSTHOG_KEY"'",
+                "api_key": "'"$POSTHOG_PAPIK"'",
                 "event": "upgrade",
                 "distinct_id": "'"$SOURCEBOT_INSTALL_ID"'",
                 "properties": {
@@ -70,23 +76,73 @@ fi
 
 echo -e "\e[34m[Info] Using config file at: '$CONFIG_PATH'.\e[0m"
 
-# Update nextjs public env variables w/o requiring a rebuild.
+# Update NextJs public env variables w/o requiring a rebuild.
 # @see: https://phase.dev/blog/nextjs-public-runtime-variables/
+{
+    # Infer NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED if it is not set
+    if [ -z "$NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED" ] && [ ! -z "$SOURCEBOT_TELEMETRY_DISABLED" ]; then
+        export NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED="$SOURCEBOT_TELEMETRY_DISABLED"
+    fi
 
-# Infer NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED if it is not set
-if [ -z "$NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED" ] && [ ! -z "$SOURCEBOT_TELEMETRY_DISABLED" ]; then
-    export NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED="$SOURCEBOT_TELEMETRY_DISABLED"
-fi
+    # Infer NEXT_PUBLIC_SOURCEBOT_VERSION if it is not set
+    if [ -z "$NEXT_PUBLIC_SOURCEBOT_VERSION" ] && [ ! -z "$SOURCEBOT_VERSION" ]; then
+        export NEXT_PUBLIC_SOURCEBOT_VERSION="$SOURCEBOT_VERSION"
+    fi
 
-# Infer NEXT_PUBLIC_SOURCEBOT_VERSION if it is not set
-if [ -z "$NEXT_PUBLIC_SOURCEBOT_VERSION" ] && [ ! -z "$SOURCEBOT_VERSION" ]; then
-    export NEXT_PUBLIC_SOURCEBOT_VERSION="$SOURCEBOT_VERSION"
-fi
+    # Always infer NEXT_PUBLIC_POSTHOG_PAPIK
+    export NEXT_PUBLIC_POSTHOG_PAPIK="$POSTHOG_PAPIK"
 
-find /app/packages/web/public /app/packages/web/.next -type f -name "*.js" |
-while read file; do
-    sed -i "s|BAKED_NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED|${NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED}|g" "$file"
-    sed -i "s|BAKED_NEXT_PUBLIC_SOURCEBOT_VERSION|${NEXT_PUBLIC_SOURCEBOT_VERSION}|g" "$file"
-done
+    # Iterate over all .js files in .next & public, making substitutions for the `BAKED_` sentinal values
+    # with their actual desired runtime value.
+    find /app/packages/web/public /app/packages/web/.next -type f -name "*.js" |
+    while read file; do
+        sed -i "s|BAKED_NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED|${NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED}|g" "$file"
+        sed -i "s|BAKED_NEXT_PUBLIC_SOURCEBOT_VERSION|${NEXT_PUBLIC_SOURCEBOT_VERSION}|g" "$file"
+        sed -i "s|BAKED_NEXT_PUBLIC_POSTHOG_PAPIK|${NEXT_PUBLIC_POSTHOG_PAPIK}|g" "$file"
+    done
+}
 
+
+# Update specifically NEXT_PUBLIC_DOMAIN_SUB_PATH w/o requiring a rebuild.
+# Ultimately, the DOMAIN_SUB_PATH sets the `basePath` param in the next.config.mjs.
+# Similar to above, we pass in a `BAKED_` sentinal value into next.config.mjs at build
+# time. Unlike above, the `basePath` configuration is set in files other than just javascript
+# code (e.g., manifest files, css files, etc.), so this section has subtle differences.
+#
+# @see: https://nextjs.org/docs/app/api-reference/next-config-js/basePath
+# @see: https://phase.dev/blog/nextjs-public-runtime-variables/
+{
+    if [ ! -z "$DOMAIN_SUB_PATH" ]; then
+        # If the sub-path is "/", this creates problems with certain replacements. For example:
+        # /BAKED_NEXT_PUBLIC_DOMAIN_SUB_PATH/_next/image -> //_next/image (notice the double slash...)
+        # To get around this, we default to an empty sub-path, which is the default when no sub-path is defined.
+        if [ "$DOMAIN_SUB_PATH" = "/" ]; then
+            DOMAIN_SUB_PATH=""
+
+        # Otherwise, we need to ensure that the sub-path starts with a slash, since this is a requirement
+        # for the basePath property. For example, assume DOMAIN_SUB_PATH=/bot, then:
+        # /BAKED_NEXT_PUBLIC_DOMAIN_SUB_PATH/_next/image -> /bot/_next/image
+        elif [[ ! "$DOMAIN_SUB_PATH" =~ ^/ ]]; then
+            DOMAIN_SUB_PATH="/$DOMAIN_SUB_PATH"
+        fi
+    fi
+
+    if [ ! -z "$DOMAIN_SUB_PATH" ]; then
+        echo -e "\e[34m[Info] DOMAIN_SUB_PATH was set to "$DOMAIN_SUB_PATH". Overriding default path.\e[0m"
+    fi
+
+    # Always set NEXT_PUBLIC_DOMAIN_SUB_PATH to DOMAIN_SUB_PATH (even if it is empty!!)
+    export NEXT_PUBLIC_DOMAIN_SUB_PATH="$DOMAIN_SUB_PATH"
+
+    # Iterate over _all_ files in the web directory, making substitutions for the `BAKED_` sentinal values
+    # with their actual desired runtime value.
+    find /app/packages/web -type f |
+    while read file; do
+        # @note: the leading "/" is required here as it is included at build time. See Dockerfile.
+        sed -i "s|/BAKED_NEXT_PUBLIC_DOMAIN_SUB_PATH|${NEXT_PUBLIC_DOMAIN_SUB_PATH}|g" "$file"
+    done
+}
+
+
+# Run supervisord
 exec supervisord -c /etc/supervisor/conf.d/supervisord.conf