+++ packages/web/src/lib/types.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { fileSourceRequestSchema, fileSourceResponseSchema, listRepositoriesResponseSchema, locationSchema, rangeSchema, repositorySchema, searchRequestSchema, searchResponseSchema } from "./schemas";
+import { fileSourceRequestSchema, fileSourceResponseSchema, listRepositoriesResponseSchema, locationSchema, rangeSchema, repositorySchema, searchRequestSchema, searchResponseSchema, symbolSchema } from "./schemas";
 
 export type KeymapType = "default" | "vim";
 
@@ -17,6 +17,8 @@ export type FileSourceResponse = z.infer<typeof fileSourceResponseSchema>;
 
 export type ListRepositoriesResponse = z.infer<typeof listRepositoriesResponseSchema>;
 export type Repository = z.infer<typeof repositorySchema>;
+
+export type Symbol = z.infer<typeof symbolSchema>;
 
 export enum SearchQueryParams {
     query = "query",