+++ packages/web/src/lib/utils.ts
@@ -3,7 +3,9 @@ import { twMerge } from "tailwind-merge"
 import githubLogo from "../../public/github.svg";
 import gitlabLogo from "../../public/gitlab.svg";
 import giteaLogo from "../../public/gitea.svg";
+import gerritLogo from "../../public/gerrit.svg";
 import { ServiceError } from "./serviceError";
+import { Repository } from "./types";
 
 export function cn(...inputs: ClassValue[]) {
     return twMerge(clsx(inputs))
@@ -30,65 +32,61 @@ export const createPathWithQueryParams = (path: string, ...queryParams: [string,
 }
 
 type CodeHostInfo = {
-    type: "github" | "gitlab" | "gitea";
-    repoName: string;
+    type: "github" | "gitlab" | "gitea" | "gerrit";
+    displayName: string;
     costHostName: string;
     repoLink: string;
     icon: string;
     iconClassName?: string;
 }
 
-export const getRepoCodeHostInfo = (repoName: string): CodeHostInfo | undefined => {
-    if (repoName.startsWith("github.com")) {
-        return {
-            type: "github",
-            repoName: repoName.substring("github.com/".length),
-            costHostName: "GitHub",
-            repoLink: `https://${repoName}`,
-            icon: githubLogo,
-            iconClassName: "dark:invert",
-        }
-    }
-    
-    if (repoName.startsWith("gitlab.com")) {
-        return {
-            type: "gitlab",
-            repoName: repoName.substring("gitlab.com/".length),
-            costHostName: "GitLab",
-            repoLink: `https://${repoName}`,
-            icon: gitlabLogo,
-        }
+export const getRepoCodeHostInfo = (repo?: Repository): CodeHostInfo | undefined => {
+    if (!repo) {
+        return undefined;
     }
 
-    if (repoName.startsWith("gitea.com")) {
-        return {
-            type: "gitea",
-            repoName: repoName.substring("gitea.com/".length),
-            costHostName: "Gitea",
-            repoLink: `https://${repoName}`,
-            icon: giteaLogo,
-        }
+    const webUrlType = repo.RawConfig ? repo.RawConfig['web-url-type'] : undefined;
+    if (!webUrlType) {
+        return undefined;
     }
 
-    return undefined;
-}
-
-export const getCodeHostFilePreviewLink = (repoName: string, filePath: string, branch: string = "HEAD"): string | undefined => {
-    const info = getRepoCodeHostInfo(repoName);
-
-    if (info?.type === "github") {
-        return `${info.repoLink}/blob/${branch}/${filePath}`;
+    const url = new URL(repo.URL);
+    const displayName = url.pathname.slice(1);
+    switch (webUrlType) {
+        case 'github':
+            return {
+                type: "github",
+                displayName: displayName,
+                costHostName: "GitHub",
+                repoLink: repo.URL,
+                icon: githubLogo,
+                iconClassName: "dark:invert",
+            }
+        case 'gitlab':
+            return {
+                type: "gitlab",
+                displayName: displayName,
+                costHostName: "GitLab",
+                repoLink: repo.URL,
+                icon: gitlabLogo,
+            }
+        case 'gitea':
+            return {
+                type: "gitea",
+                displayName: displayName,
+                costHostName: "Gitea",
+                repoLink: repo.URL,
+                icon: giteaLogo,
+            }
+        case 'gitiles':
+            return {
+                type: "gerrit",
+                displayName: displayName,
+                costHostName: "Gerrit",
+                repoLink: repo.URL,
+                icon: gerritLogo,
+            }
     }
-
-    if (info?.type === "gitlab") {
-        return `${info.repoLink}/-/blob/${branch}/${filePath}`;
-    }
-
-    if (info?.type === "gitea") {
-        return `${info.repoLink}/src/branch/${branch}/${filePath}`;
-    }
-
-    return undefined;
 }
 
 export const isServiceError = (data: unknown): data is ServiceError => {
@@ -99,7 +97,7 @@ export const isServiceError = (data: unknown): data is ServiceError => {
         'message' in data;
 }
 
-export const getEnv = (env: string | undefined, defaultValue = '') => {
+export const getEnv = (env: string | undefined, defaultValue?: string) => {
 	return env ?? defaultValue;
 }
 
@@ -118,4 +116,9 @@ export const getEnvBoolean = (env: string | undefined, defaultValue: boolean) =>
 export const base64Decode = (base64: string): string => {
     const binString = atob(base64);
     return Buffer.from(Uint8Array.from(binString, (m) => m.codePointAt(0)!).buffer).toString();
+}
+
+// @see: https://stackoverflow.com/a/65959350/23221295
+export const isDefined = <T>(arg: T | null | undefined): arg is T extends null | undefined ? never : T => {
+    return arg !== null && arg !== undefined;
 }